requirements.txt
pandas
sqlite3
python-dotenv
groq
streamlit
schedule
smtplib
